Backend/app/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
pydantic
fastapi-users[sqlalchemy]
python-multipart
requests
motor # for async mongo db
httpx # async HTTP requests
python-jose # (for JWT)
